requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
anyio=3.6.2=pyhd8ed1ab_0
appdirs=1.4.4=pypi_0
argon2-cffi=21.3.0=pyhd8ed1ab_0
argon2-cffi-bindings=21.2.0=py310h5764c6d_2
astropy=5.1.1=pypi_0
asttokens=2.0.8=pyhd8ed1ab_0
attrs=22.1.0=pyh71513ae_1
babel=2.10.3=pyhd8ed1ab_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
beautifulsoup4=4.11.1=pyha770c72_0
bleach=5.0.1=pyhd8ed1ab_0
bokeh=2.4.3=py_0
bqplot=0.12.36=pypi_0
brotlipy=0.7.0=py310h5764c6d_1004
bzip2=1.0.8=h7f98852_4
ca-certificates=2022.9.24=ha878542_0
certifi=2022.9.24=pyhd8ed1ab_0
cffi=1.15.1=py310h255011f_1
charset-normalizer=2.1.1=pyhd8ed1ab_0
colorama=0.4.5=pyhd8ed1ab_0
colorcet=3.0.1=pypi_0
cryptography=38.0.2=py310h597c629_0
debugpy=1.6.3=py310hd8f1fbe_0
decorator=5.1.1=pyhd8ed1ab_0
defusedxml=0.7.1=pyhd8ed1ab_0
entrypoints=0.4=pyhd8ed1ab_0
executing=1.1.1=pyhd8ed1ab_0
flit-core=3.7.1=pyhd8ed1ab_0
freetype=2.12.1=hca18f0e_0
future=0.18.2=pypi_0
holoviews=1.15.2=pypi_0
hrv-analysis=1.0.4=pypi_0
idna=3.4=pyhd8ed1ab_0
importlib-metadata=4.11.4=py310hff52083_0
importlib_resources=5.10.0=pyhd8ed1ab_0
ipyfilechooser=0.6.0=pypi_0
ipykernel=6.16.0=pyh210e3f2_0
ipython=8.5.0=pyh41d4057_1
ipython_genutils=0.2.0=py_1
ipywidgets=8.0.2=pypi_0
jedi=0.18.1=pyhd8ed1ab_2
jinja2=3.1.2=pyhd8ed1ab_1
jpeg=9e=h166bdaf_2
json5=0.9.5=pyh9f0ad1d_0
jsonschema=4.16.0=pyhd8ed1ab_0
jupyter-bokeh=3.0.5=pypi_0
jupyter_client=7.4.2=pyhd8ed1ab_0
jupyter_core=4.11.1=py310hff52083_0
jupyter_server=1.21.0=pyhd8ed1ab_0
jupyterlab=3.4.8=pyhd8ed1ab_0
jupyterlab-widgets=3.0.3=pypi_0
jupyterlab_pygments=0.2.2=pyhd8ed1ab_0
jupyterlab_server=2.16.0=pyhd8ed1ab_0
lcms2=2.12=hddcbb42_0
ld_impl_linux-64=2.39=hc81fddc_0
lerc=4.0.0=h27087fc_0
libblas=3.9.0=16_linux64_openblas
libcblas=3.9.0=16_linux64_openblas
libdeflate=1.14=h166bdaf_0
libffi=3.4.2=h7f98852_5
libgcc-ng=12.2.0=h65d4601_18
libgfortran-ng=12.2.0=h69a702a_18
libgfortran5=12.2.0=h337968e_18
libgomp=12.2.0=h65d4601_18
liblapack=3.9.0=16_linux64_openblas
libnsl=2.0.0=h7f98852_0
libopenblas=0.3.21=pthreads_h78a6416_3
libpng=1.6.38=h753d276_0
libsodium=1.0.18=h36c2ea0_1
libsqlite=3.39.4=h753d276_0
libstdcxx-ng=12.2.0=h46fd767_18
libtiff=4.4.0=h55922b4_4
libuuid=2.32.1=h7f98852_1000
libwebp-base=1.2.4=h166bdaf_0
libxcb=1.13=h7f98852_1004
libzlib=1.2.13=h166bdaf_4
markdown=3.4.1=pyhd8ed1ab_0
markupsafe=2.1.1=py310h5764c6d_1
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mistune=2.0.4=pyhd8ed1ab_0
mne=1.2.1=pypi_0
nbclassic=0.4.5=pyhd8ed1ab_0
nbclient=0.7.0=pyhd8ed1ab_0
nbconvert=7.2.1=pyhd8ed1ab_0
nbconvert-core=7.2.1=pyhd8ed1ab_0
nbconvert-pandoc=7.2.1=pyhd8ed1ab_0
nbformat=5.7.0=pyhd8ed1ab_0
ncurses=6.3=h27087fc_1
nest-asyncio=1.5.6=pyhd8ed1ab_0
nolds=0.5.2=pypi_0
notebook=6.5.1=pyha770c72_0
notebook-shim=0.2.0=pyhd8ed1ab_0
numpy=1.23.4=py310h53a5b5f_0
openjpeg=2.5.0=h7d73246_1
openssl=1.1.1q=h166bdaf_0
packaging=21.3=pyhd8ed1ab_0
pandoc=2.19.2=ha770c72_0
pandocfilters=1.5.0=pyhd8ed1ab_0
panel=0.14.0=py_0
param=1.12.2=py_0
parso=0.8.3=pyhd8ed1ab_0
pexpect=4.8.0=pyh9f0ad1d_2
pickleshare=0.7.5=py_1003
pillow=9.2.0=py310hbd86126_2
pip=22.3=pyhd8ed1ab_0
pkgutil-resolve-name=1.3.10=pyhd8ed1ab_0
pooch=1.6.0=pypi_0
prometheus_client=0.15.0=pyhd8ed1ab_0
prompt-toolkit=3.0.31=pyha770c72_0
psutil=5.9.3=py310h5764c6d_0
pthread-stubs=0.4=h36c2ea0_1001
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pycparser=2.21=pyhd8ed1ab_0
pyct=0.4.8=py_0
pyct-core=0.4.8=py_0
pyedflib=0.1.30=pypi_0
pyerfa=2.0.0.1=pypi_0
pygments=2.13.0=pyhd8ed1ab_0
pyopenssl=22.1.0=pyhd8ed1ab_0
pyparsing=3.0.9=pyhd8ed1ab_0
pyqt5=5.15.7=pypi_0
pyqt5-qt5=5.15.2=pypi_0
pyqt5-sip=12.11.0=pypi_0
pyrsistent=0.18.1=py310h5764c6d_1
pysocks=1.7.1=pyha2e5f31_6
python=3.10.6=h582c2e5_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python-fastjsonschema=2.16.2=pyhd8ed1ab_0
python_abi=3.10=2_cp310
pytz=2022.5=pyhd8ed1ab_0
pyviz_comms=2.2.1=py_0
pyyaml=6.0=py310h5764c6d_4
pyzmq=24.0.1=py310h330234f_0
readline=8.1.2=h0f457ee_0
requests=2.28.1=pyhd8ed1ab_1
send2trash=1.8.0=pyhd8ed1ab_0
setuptools=65.5.0=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
sniffio=1.3.0=pyhd8ed1ab_0
soundfile=0.11.0=pypi_0
soupsieve=2.3.2.post1=pyhd8ed1ab_0
stack_data=0.5.1=pyhd8ed1ab_0
terminado=0.16.0=pyh41d4057_0
tinycss2=1.2.1=pyhd8ed1ab_0
tk=8.6.12=h27826a3_0
tomli=2.0.1=pyhd8ed1ab_0
tornado=6.2=py310h5764c6d_0
tqdm=4.64.1=pyhd8ed1ab_0
traitlets=5.5.0=pyhd8ed1ab_0
traittypes=0.2.1=pypi_0
typing_extensions=4.4.0=pyha770c72_0
tzdata=2022e=h191b570_0
urllib3=1.26.11=pyhd8ed1ab_0
wcwidth=0.2.5=pyh9f0ad1d_2
webencodings=0.5.1=py_1
websocket-client=1.4.1=pyhd8ed1ab_0
wfdb=4.0.0=pypi_0
wheel=0.37.1=pyhd8ed1ab_0
widgetsnbextension=4.0.3=pypi_0
xorg-libxau=1.0.9=h7f98852_0
xorg-libxdmcp=1.1.3=h7f98852_0
xz=5.2.6=h166bdaf_0
yaml=0.2.5=h7f98852_2
zeromq=4.3.4=h9c3ff4c_1
zipp=3.9.0=pyhd8ed1ab_0
zstd=1.5.2=h6239696_4
